+++ android/app/src/main/kotlin/com/b1zz1/template/flutter_auth/flutter_auth/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.login_app
+package com.b1zz1.template.flutter_auth.flutter_auth
 
 import io.flutter.embedding.android.FlutterActivity
