+++ pubspec.yaml
@@ -1,5 +1,5 @@
-name: login_app
-description: "Application allowing sign in with Google."
+name: flutter_auth
+description: "Flutter application with Google and Apple authentication"
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
@@ -60,7 +60,8 @@ flutter:
 
   # To add assets to your application, add an assets section, like this:
   assets:
-    - assets/images/
+      - assets/images/
+  #   - images/a_dot_burr.jpeg
   #   - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
